+++ src/models/qwen_embedding.py
@@ -27,7 +27,7 @@ class Qwen3Embedding(BaseModel):
         model_id = 'Qwen/Qwen3-Embedding-' + size
         self.device = device
         self.tokenizer = AutoTokenizer.from_pretrained(model_id, trust_remote_code=True)
-        self.model = AutoModel.from_pretrained(model_id, trust_remote_code=True).to(device, dtype=torch.bfloat16)
+        self.model = AutoModel.from_pretrained(model_id, trust_remote_code=True).to(device, dtype=torch.float32)
         self.prompt_query = 'Given a natural language query, retrieve formal Coq statements whose docstrings best match the intent of the query.'
     
     def generate(self, sentence:str, query=False) -> Tensor:
